LPS-31740 Fix duplicate id
The aui:input checkbox for "Automatically apply changes done to the page template" has the same name (and thus by default the same id) in the page details view and in the add page dialog. Since the latest is included (though hidden) in the html of the first, in order to prevent JS errors in actions on this checkbox, we have to explicitly assign it a different id in the add_layout.jsp, as done in other inputs of this view.