LPS-45070 Ignore the modified date of the custom RTL CSS file when generating and serving the CSS cache resource
The CSS cache resource is generated by combining the main CSS resource with the custom RTL CSS file, if it exists. Since the modifiedDate can only be equals to one of them, it's been decided to use that of the main CSS resource. This means that if the custom RTL CSS is modified afterwards, the previous RTL CSS cache resource must be removed to enforce its regeneration. This will be added to documentation.