LPS-47036 call updateAsset when cascading file entry types
Call DLAppHelperLocalService.updateAsset when propagating changes in the folder entry type.
This code is called from DLApp*Service.addFolder, but we didn't add the call to dlAppHelper
in there because it could be highly inefficient: we would traverse the tree of files twice;
also, from DLApp*Service there is no way to know if the folder allows type overrides or not,
or if the file entry needs to be updated (i.e. its file entry type may be already the correct
one).