LPS-47465 Add more convenience methods with sensible defaults.
These methods allow to create both Object and List ServiceTrackerMaps
specifying custom mappers and/or comparators. Alas not all combinations
can be properly enclosed in a factory and this will likely be a
candidate for a builder in the future but, for the moment, it is
sufficiently handy to have this factory.