LPS-47465 Add tuple class
This class delegates both hash and equals to the contained service
reference. It also delegates compareTo to the contained service
reference, so it will behave as a key exactly the same as a service
reference used to behave.

This way the buckets can easily store these tuples ordered by their
references and obtain the associated services.