LPS-51358 test WF when adding a file entry
For testing the workflow I introduce here a testing pattern based on the
RAII idiom of C++.

In some tests there is a need to slightly modify the environment (in
this particular case, I wanted to intercept calls to WF handlers to
count the number of invocations). Making sure that the env. is modified
and cleaned up after the test using the common Java idioms (try /
finally) lends to very verbose tests.

I've taken advantage of Java 7 try-with-resources syntax for this
purpose. The idea is to encapsulate this environment changes in an
object, where the state is initialized in the constructor and cleaned up
in the `close` method. The try block semantics will guarantee that
everything works correctly (calling the close method automatically) and
all the logic will be encapsulated in an object, making it reusable
across different tests.