LPS-52140 Pure global variables are evil
Prefer adding variables in Liferay namespace instead to create
more global variables. themeDisplay (which pattern was followed here)
was created probably for performance issues in the past when accessing
properties like this - "prop1.prop2.prop3" was slow.
For many years it is not anymore and adding global variables only polutes
the global namespace, nothing else.