LPS-53795 Add basic extension models for the panel
The user personal panel will consist of:

  - `PanelCategory` instances , which will act as a container for other
    `PanelEntry` instances. Categories will support an arbirtrary
    nesting depth, and will represent "sections" in the panel.

  - `PanelApp` instances, which will represent the applications visible
    in the panel.

To register a new panel entry (category or app), it is only required to
implement one of the supported interfaces, and to define the bundle
property `panel.category`. This property will indicate the parent of the
entry. A special root category is provided (`panel.category=root`).