LPS-53400 Change WikiGroupServiceSettings to the new portal-settings format (note that we don't need the implementation any more as it is provided automagically by portal-settings).
The remaining extra methods that were not part of the configuration, but extensions to it, have been moved to a pseudo-mixin class (WikiGroupServiceSettingsExtraImpl) that will be used by portal-settings to create the WikiGroupServiceSettings proxy. We declare the extra methods in an interface to be able to force WikiGroupServiceSettingsExtraImpl to implement it and don't forget any new method we may add in the future.